исправленная структура задания
